feat: move noop client to a separate package
