+++ pkg/client/noop/noop_client.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package client
+package noop
 
 import (
 	"context"
@@ -18,20 +18,20 @@ import (
 
 const componentNoopName = "noop"
 
-// NoopClient is an implementation of Output that discards all records
+// Client is an implementation of Output that discards all records
 // but keeps metrics and increments counters
-type NoopClient struct {
+type Client struct {
 	ctx      context.Context
 	logger   logr.Logger
 	endpoint string
 	metrics  *metrics.FluentBitGardenerMetrics
 }
 
-var _ api.Output = &NoopClient{}
+var _ api.Output = &Client{}
 
-// NewNoopClient creates a new NoopClient that discards all records
-func NewNoopClient(ctx context.Context, cfg config.Config, logger logr.Logger, m *metrics.FluentBitGardenerMetrics) (api.Output, error) {
-	client := &NoopClient{
+// New creates a new NoopClient that discards all records
+func New(ctx context.Context, cfg config.Config, logger logr.Logger, m *metrics.FluentBitGardenerMetrics) (api.Output, error) {
+	client := &Client{
 		ctx:      ctx,
 		endpoint: cfg.OTLPConfig.Endpoint,
 		logger:   logger.WithValues("endpoint", cfg.OTLPConfig.Endpoint),
@@ -44,7 +44,7 @@ func NewNoopClient(ctx context.Context, cfg config.Config, logger logr.Logger, m
 }
 
 // Handle processes and discards the log entry while incrementing metrics
-func (c *NoopClient) Handle(_ types.OutputEntry) error {
+func (c *Client) Handle(_ types.OutputEntry) error {
 	// Increment the dropped logs counter since we're discarding the record
 	c.metrics.DroppedLogs.WithLabelValues(c.endpoint, "noop").Inc()
 
@@ -53,16 +53,16 @@ func (c *NoopClient) Handle(_ types.OutputEntry) error {
 }
 
 // Stop shuts down the client immediately
-func (c *NoopClient) Stop() {
+func (c *Client) Stop() {
 	c.logger.V(2).Info(fmt.Sprintf("stopping %s", componentNoopName))
 }
 
 // StopWait stops the client - since this is a no-op client, it's the same as Stop
-func (c *NoopClient) StopWait() {
+func (c *Client) StopWait() {
 	c.logger.V(2).Info(fmt.Sprintf("stopping %s with wait", componentNoopName))
 }
 
 // GetEndpoint returns the configured endpoint
-func (c *NoopClient) GetEndpoint() string {
+func (c *Client) GetEndpoint() string {
 	return c.endpoint
 }
